+++ nienowt/controllers/humans/humans.js
@@ -11,18 +11,11 @@
       if (this.editShow !== 'new') return this.editShow = 'new';
       this.editShow = name;
     };
-    this.confirmChange = function(human, buttonName, curHuman){
-      if (!this.editConfirmation) return this.editConfirmation = true;
-      if(buttonName === 'delete') return this.removeHuman(human);
-      if(buttonName === 'edit') return this.editHuman(human, curHuman);
-    };
 
     this.getHumans = function() {
       $http.get(mainRoute)
       .then((results) => {
-        console.log(results);
         this.humans = results.data;
-        console.log(results.data);
       },(err) => {
         if (err) console.log(err);
       });
@@ -34,14 +27,17 @@
         this.newHuman = {};
       });
     };
-    this.editHuman = function(changedHuman, human){
-      $http.put(mainRoute + '/' + human._id, changedHuman)
+    this.editHuman = function(human){
+      if (!this.editConfirmation) return this.editConfirmation = true;
+      $http.put(mainRoute + '/' + human._id, this.changedHuman)
       .then(() => {
         this.humans = this.humans.filter((g) => g._id != human._id);
-        this.humans.push(changedHuman);
+        this.humans.push(this.changedHuman);
+        this.editShow = 'new';
       });
     };
     this.removeHuman = function(human){
+      if (!this.editConfirmation) return this.editConfirmation = true;
       $http.delete(mainRoute + '/' + human._id)
       .then(() => {
         this.humans = this.humans.filter((g) => g._id != human._id);
@@ -51,7 +47,7 @@
     this.reset = function(){
       this.editShow = 'new';
       this.changedHuman = {};
-      this.editConfirmation = false;
+      return this.editConfirmation = false;
     };
   }]);
 })()
